deleted uf3/optimize.py and bug fixes to uf3/util/optimize.py
